ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pdfplumber==0.11.4
python-docx==1.1.2
spacy==3.7.6
sentence-transformers==3.0.1
numpy==1.26.4
python-multipart==0.0.9
scikit-learn==1.5.1
